+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+- Bump svelte minimum version to `4.2.1`
 - Bump esbuild minimum version to `0.17.0`
 - Remove `overzealous` cache option (was unusable)